UPDATES 1. Added different players 2. Added weights as a result of training
