+++ raise_player.py
@@ -39,7 +39,7 @@ class Group18Player(BasePokerPlayer):
 
             model = Model(inputs=[input_cards, input_actions,input_position], outputs=out)
             if self.vvh == 0:
-                model.load_weights('setup/weights.h5', by_name=True)
+                model.load_weights('setup/initial_weights.h5', by_name=True)
 
             model.compile(optimizer='rmsprop', loss='mse')
