+++ reinforced_player.py
@@ -1,0 +1,34 @@
+from pypokerengine.players import BasePokerPlayer
+from pypokerengine.api.emulator import Emulator
+from pypokerengine.utils.game_state_utils import restore_game_state
+
+from raise_player import Group18Player
+
+class RLPLayer(BasePokerPlayer):
+
+    # Setup Emulator object by registering game information
+    def receive_game_start_message(self, game_info):
+        player_num = game_info["player_num"]
+        max_round = game_info["rule"]["max_round"]
+        small_blind_amount = game_info["rule"]["small_blind_amount"]
+        ante_amount = game_info["rule"]["ante"]
+        blind_structure = game_info["rule"]["blind_structure"]
+
+        self.emulator = Emulator()
+        self.emulator.set_game_rule(player_num, max_round, small_blind_amount, ante_amount)
+        self.emulator.set_blind_structure(blind_structure)
+
+        # Register algorithm of each player which used in the simulation.
+        for player_info in game_info["seats"]["players"]:
+            self.emulator.register_player(player_info["uuid"], Group18Player())
+
+    def declare_action(self, valid_actions, hole_card, round_state):
+        game_state = restore_game_state(round_state)
+        # decide action by using some simulation result
+        # updated_state, events = self.emulator.apply_action(game_state, "fold")
+        updated_state, events = self.emulator.run_until_round_finish(game_state)
+        # updated_state, events = self.emulator.run_until_game_finish(game_state)
+        if self.is_good_simulation_result(updated_state):
+            return # you would declare CALL or RAISE action
+        else:
+            return "fold", 0
