+++ setup/halfway_honest_weights_1.h5
